ELTA 2.0/requirements.txt
einops==0.4.1
matplotlib==3.3.4
nni==2.6.1
numpy==1.19.5
pandas==1.1.5
Pillow==10.2.0
scikit_learn==1.4.0
scipy==1.5.4
timm==0.6.12
torch==1.10.1
torchvision==0.11.2
tqdm==4.64.1
